Use context manager for repository handling
